% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveols.R
\name{solveols}
\alias{solveols}
\title{Solve linear system}
\usage{
solveols(A, b, iter, c = 1, type = "GS")
}
\arguments{
\item{A}{Matrix.}

\item{b}{Vector. Its dimension should match that of matrix \code{A}.}

\item{iter}{Number of iterations in the calculation.}

\item{c}{Number of cores used in parallel computation.}

\item{type}{Type of iteration method, 'GS' means Gauss-Seidel and "Jacobi" for Jacobi}
}
\value{
The solution vector \code{x} for linear system \code{Ax = b}.
}
\description{
Solve linear system \code{Ax = b} using iteration methods Jacobi, Gauss Seidl and parallel Jacobi.
}
